Split runner process helpers
